[Fix] Bug in battery.sh

+++ src/battery.sh
@@ -5,14 +5,16 @@ SLEEP=2
 
 while true; do
     BAT_PERCENT=$(acpi | grep -o "[0-9]*%" | grep -o "[0-9]*")
-    if [[ $BAT_PERCENT -le $DANGER ]]; then
-        i3-nagbar -m "Low Battery!!!" & PID=$!
-    fi
-    sleep $SLEEP"m"
+    if [[ -z $(acpi | grep Charging) ]]; then
+        if [[ $BAT_PERCENT -le $DANGER ]]; then
+            i3-nagbar -m "Low Battery!!!" & PID=$!
+        fi
+        sleep $SLEEP"m"
 
-    if [[ -n $(acpi | grep "Discharging") ]]; then
-        echo shit
-    else
-        kill -9 $PID
+        if [[ -n $(acpi | grep "Discharging") ]]; then
+            sleep 0
+        else
+            kill -9 $PID
+        fi
     fi
 done
